add NativePeripheralManagerCallbacks
