+++ Packages/com.teach310.core-bluetooth-for-unity/Runtime/NativeCallbacks/NativePeripheralManagerCallbacks.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1e86702b2f853478cb8ab72e3e55c33a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
