+++ Packages/com.teach310.core-bluetooth-for-unity/Runtime/NativePeripheralManagerProxy.cs
@@ -5,21 +5,25 @@ using System.Runtime.InteropServices;
 
 namespace CoreBluetooth
 {
+    internal interface INativePeripheralManagerCallbacks
+    {
+        void Register(SafeNativePeripheralManagerHandle handle, INativePeripheralManagerDelegate peripheralManagerDelegate);
+        void Unregister(SafeNativePeripheralManagerHandle handle);
+    }
+
     internal class NativePeripheralManagerProxy : IDisposable
     {
-        static Dictionary<IntPtr, INativePeripheralManagerDelegate> s_peripheralManagerDelegateMap = new Dictionary<IntPtr, INativePeripheralManagerDelegate>();
-
         readonly SafeNativePeripheralManagerHandle _handle;
+        readonly INativePeripheralManagerCallbacks _callbacks;
 
         internal NativePeripheralManagerProxy(SafeNativePeripheralManagerHandle handle, INativePeripheralManagerDelegate peripheralManagerDelegate)
         {
             _handle = handle;
-            if (peripheralManagerDelegate != null)
+            _callbacks = ServiceLocator.Instance.Resolve<INativePeripheralManagerCallbacks>();
+            if (_callbacks != null)
             {
-                s_peripheralManagerDelegateMap[handle.DangerousGetHandle()] = peripheralManagerDelegate;
+                _callbacks.Register(handle, peripheralManagerDelegate);
             }
-
-            RegisterHandlers();
         }
 
         internal void AddService(SafeNativeMutableServiceHandle service)
@@ -100,92 +104,9 @@ namespace CoreBluetooth
             NativeMethods.cb4u_peripheral_manager_respond_to_request(_handle, request.Handle, (int)result);
         }
 
-        void RegisterHandlers()
-        {
-            NativeMethods.cb4u_peripheral_manager_register_handlers(
-                _handle,
-                DidUpdateState,
-                DidAddService,
-                DidStartAdvertising,
-                DidSubscribeToCharacteristic,
-                DidUnsubscribeFromCharacteristic,
-                IsReadyToUpdateSubscribers,
-                DidReceiveReadRequest,
-                DidReceiveWriteRequests
-            );
-        }
-
         public void Dispose()
         {
-            s_peripheralManagerDelegateMap.Remove(_handle.DangerousGetHandle());
-        }
-
-        static INativePeripheralManagerDelegate GetDelegate(IntPtr peripheralPtr)
-        {
-            return s_peripheralManagerDelegateMap.GetValueOrDefault(peripheralPtr);
-        }
-
-        [AOT.MonoPInvokeCallback(typeof(NativeMethods.CB4UPeripheralManagerDidUpdateStateHandler))]
-        public static void DidUpdateState(IntPtr peripheralPtr, CBManagerState state)
-        {
-            GetDelegate(peripheralPtr)?.DidUpdateState(state);
-        }
-
-        [AOT.MonoPInvokeCallback(typeof(NativeMethods.CB4UPeripheralManagerDidAddServiceHandler))]
-        public static void DidAddService(IntPtr peripheralPtr, IntPtr serviceUUIDPtr, int errorCode)
-        {
-            GetDelegate(peripheralPtr)?.DidAddService(
-                Marshal.PtrToStringUTF8(serviceUUIDPtr),
-                CBError.CreateOrNullFromCode(errorCode)
-            );
-        }
-
-        [AOT.MonoPInvokeCallback(typeof(NativeMethods.CB4UPeripheralManagerDidStartAdvertisingHandler))]
-        public static void DidStartAdvertising(IntPtr peripheralPtr, int errorCode)
-        {
-            GetDelegate(peripheralPtr)?.DidStartAdvertising(CBError.CreateOrNullFromCode(errorCode));
-        }
-
-        [AOT.MonoPInvokeCallback(typeof(NativeMethods.CB4UPeripheralManagerDidSubscribeToCharacteristicHandler))]
-        public static void DidSubscribeToCharacteristic(IntPtr peripheralPtr, IntPtr centralPtr, IntPtr serviceUUIDPtr, IntPtr characteristicUUIDPtr)
-        {
-            GetDelegate(peripheralPtr)?.DidSubscribeToCharacteristic(
-                new SafeNativeCentralHandle(centralPtr),
-                Marshal.PtrToStringUTF8(serviceUUIDPtr),
-                Marshal.PtrToStringUTF8(characteristicUUIDPtr)
-            );
-        }
-
-        [AOT.MonoPInvokeCallback(typeof(NativeMethods.CB4UPeripheralManagerDidUnsubscribeFromCharacteristicHandler))]
-        public static void DidUnsubscribeFromCharacteristic(IntPtr peripheralPtr, IntPtr centralPtr, IntPtr serviceUUIDPtr, IntPtr characteristicUUIDPtr)
-        {
-            GetDelegate(peripheralPtr)?.DidUnsubscribeFromCharacteristic(
-                new SafeNativeCentralHandle(centralPtr),
-                Marshal.PtrToStringUTF8(serviceUUIDPtr),
-                Marshal.PtrToStringUTF8(characteristicUUIDPtr)
-            );
-        }
-
-        [AOT.MonoPInvokeCallback(typeof(NativeMethods.CB4UPeripheralManagerIsReadyToUpdateSubscribersHandler))]
-        public static void IsReadyToUpdateSubscribers(IntPtr peripheralPtr)
-        {
-            GetDelegate(peripheralPtr)?.IsReadyToUpdateSubscribers();
-        }
-
-        [AOT.MonoPInvokeCallback(typeof(NativeMethods.CB4UPeripheralManagerDidReceiveReadRequestHandler))]
-        public static void DidReceiveReadRequest(IntPtr peripheralPtr, IntPtr requestPtr)
-        {
-            GetDelegate(peripheralPtr)?.DidReceiveReadRequest(
-                new SafeNativeATTRequestHandle(requestPtr)
-            );
-        }
-
-        [AOT.MonoPInvokeCallback(typeof(NativeMethods.CB4UPeripheralManagerDidReceiveWriteRequestsHandler))]
-        public static void DidReceiveWriteRequests(IntPtr peripheralPtr, IntPtr requestsPtr)
-        {
-            GetDelegate(peripheralPtr)?.DidReceiveWriteRequests(
-                new SafeNativeATTRequestsHandle(requestsPtr)
-            );
+            _callbacks?.Unregister(_handle);
         }
     }
 }
